requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
channels>=4.0.0
channels-redis>=4.0.0
firebase-admin>=6.0.0
cryptography>=40.0.0
python-dotenv>=0.21.0
asgiref>=3.5.0
pyrebase4>=4.6.0
daphne>=4.0.0